contactForm component and adding react toasts
