+++ components/ContactForm/index.ts
@@ -1,0 +1,3 @@
+import ContactForm from "./component";
+
+export default ContactForm;
