+++ components/ContactForm/styles.module.scss
@@ -1,11 +1,6 @@
 //IMPORTS
-@import '../base/variables';
-@import '../function/mixins';
-
-.contactPage{
-    min-height: 100vh;
-    width: 100%;
-}
+@import '../../styles/base/variables';
+@import '../../styles/function/mixins';
 
 .form{
     width: 500px;
