+++ package.json
@@ -14,6 +14,7 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-hook-form": "^7.43.1",
+    "react-hot-toast": "^2.4.0",
     "sass": "^1.58.0",
     "zod": "^3.20.6"
   },
