+++ pages/_app.tsx
@@ -1,5 +1,7 @@
 //COMPONENTS
 import Navbar from '../components/Navbar';
+//LIBRARIES
+import { Toaster } from 'react-hot-toast';
 //STYLES  
 import '../styles/App.scss';
 
@@ -8,6 +10,7 @@ function MyApp({ Component, pageProps }) {
     <>
       <Navbar />
       <Component {...pageProps} />
+      <Toaster />
     </>
   )
 }
